getPostsPreviewWithCategoryName defined as a global constant
