+++ vite-project-kabinet/src/Content/PaginatedList.jsx
@@ -5,6 +5,7 @@ import { useEffect } from "react";
 import { PaginationNav } from "./PaginationNav.jsx";
 import { BASE_API_URL } from "../constants/baseApiUrl.js";
 import { stripHTMLTags } from "../utils/stripHtmlTags.js";
+import { getPostsPreviewWithCategoryName } from "../constants/getPostsPreview.js";
 
 export const PaginatedList = ({itemsPerPage}) => {
     const [currentPage, setCurrentPage] = useState(1);
@@ -17,50 +18,51 @@ export const PaginatedList = ({itemsPerPage}) => {
         }
     };
 
-    const getPostsWithCategoryName = async (perPage, page) => {
+    getPostsPreviewWithCategoryName(itemsPerPage, currentPage, setTotalPages, BASE_API_URL, stripHTMLTags)
+    // const getPostsWithCategoryName = async (perPage, page) => {
 
-        const reqPosts = await fetch(`${BASE_API_URL}/posts?per_page=${perPage}&page=${page}`)
-        const totalPages = reqPosts.headers.get("X-WP-TotalPages");
-        setTotalPages(totalPages)
-        const posts = await reqPosts.json()
+    //     const reqPosts = await fetch(`${BASE_API_URL}/posts?per_page=${perPage}&page=${page}`)
+    //     const totalPages = reqPosts.headers.get("X-WP-TotalPages");
+    //     setTotalPages(totalPages)
+    //     const posts = await reqPosts.json()
     
-        const reqCategories = await fetch(`${BASE_API_URL}/categories?_fields=id,name`)
-        const categories = await reqCategories.json()
+    //     const reqCategories = await fetch(`${BASE_API_URL}/categories?_fields=id,name`)
+    //     const categories = await reqCategories.json()
     
-        //chcem získať pole objektov, kde bude každý objekt obsahovať aj konkrétny názov kategórie
+    //     //chcem získať pole objektov, kde bude každý objekt obsahovať aj konkrétny názov kategórie
         
-        const updatedPosts = posts.map((post) => {
-            const postCategsArray = post.categories
-            const newPostCategsArray = postCategsArray.map((catID) => {
-                const categObject = categories.find((category) => category.id === catID )
-                return categObject ? categObject.name : "category not found"
-            })
+    //     const updatedPosts = posts.map((post) => {
+    //         const postCategsArray = post.categories
+    //         const newPostCategsArray = postCategsArray.map((catID) => {
+    //             const categObject = categories.find((category) => category.id === catID )
+    //             return categObject ? categObject.name : "category not found"
+    //         })
     
-            // Aby program správne fungoval, každý článok musí obsahovať vždy práve jednu kategóriu.
-            // Kabinet má aktuálne pár článkov takých, kde sú pridelené 2 kategórie
+    //         // Aby program správne fungoval, každý článok musí obsahovať vždy práve jednu kategóriu.
+    //         // Kabinet má aktuálne pár článkov takých, kde sú pridelené 2 kategórie
     
-            // todo: Dá sa nastaviť WP tak, aby uživateľ mohol zakliknúť vždy len jednu kategóriu ?
-            // Nedá. Muselo by se vytvořit nové pole, které by fungovalo stejně jen s tímhle omezením. Znamenalo by to, že by se musely upravit hodnoty pro všechny články.
+    //         // todo: Dá sa nastaviť WP tak, aby uživateľ mohol zakliknúť vždy len jednu kategóriu ?
+    //         // Nedá. Muselo by se vytvořit nové pole, které by fungovalo stejně jen s tímhle omezením. Znamenalo by to, že by se musely upravit hodnoty pro všechny články.
 
-            // todo: pridať žánre vo WP
-            // todo: nastaviť WP tak, aby šlo pridávať fotky
-            // todo: nastaviť WP tak, aby mal autor svoju vlastnú kolonku
+    //         // todo: pridať žánre vo WP
+    //         // todo: nastaviť WP tak, aby šlo pridávať fotky
+    //         // todo: nastaviť WP tak, aby mal autor svoju vlastnú kolonku
     
-            return {
-                author: "Název autora",
-                perex: stripHTMLTags(post.excerpt.rendered),
-                id: post.id,
-                image: "",
-                title: post.title.rendered,
-                genre: newPostCategsArray.join(),                
-            }
-        })
+    //         return {
+    //             author: "Název autora",
+    //             perex: stripHTMLTags(post.excerpt.rendered),
+    //             id: post.id,
+    //             image: "",
+    //             title: post.title.rendered,
+    //             genre: newPostCategsArray.join(),                
+    //         }
+    //     })
     
-        return updatedPosts
-    }
+    //     return updatedPosts
+    // }
 
     useEffect(() => {
-        getPostsWithCategoryName(itemsPerPage, currentPage).then((data) => {
+        getPostsPreviewWithCategoryName(itemsPerPage, currentPage, setTotalPages, BASE_API_URL, stripHTMLTags).then((data) => {
             setItems(data)
         })
     }, [currentPage])
